# src/api_service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API с JWT авторизацией",
    "description": "Пример API, использующего JWT для авторизации в Swagger UI",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/cdn/upload": {
      "post": {
        "tags": [
          "CDN"
        ],
        "summary": "Upload Image",
        "operationId": "upload_image_v1_cdn_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_v1_cdn_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Image uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageLinkDTO"
                },
                "example": {
                  "id": "b1d0d3c1-9a7d-4b7b-9d8f-6e6b1f5b5b8b",
                  "path": "/v1/cdn/file/b1d0d3c1-9a7d-4b7b-9d8f-6e6b1f5b5b8b.jpg"
                }
              }
            }
          },
          "400": {
            "description": "Invalid image data",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid image data"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/cdn/file/{filename}": {
      "get": {
        "tags": [
          "CDN"
        ],
        "summary": "Get Image",
        "description": "Get file from cdn",
        "operationId": "get_image_v1_cdn_file__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File found",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/jpeg": {
                "example": "image/jpeg"
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "File not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_image_v1_cdn_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_v1_cdn_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageLinkDTO": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "ImageLinkDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен для авторизации. Формат: Bearer <токен>"
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}